moved module load warning
