+++ spatialprofilingtoolbox/control/scripts/configure.py
@@ -13,7 +13,7 @@ import importlib.resources
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     from spatialprofilingtoolbox import get_workflow
     from spatialprofilingtoolbox import get_workflow_names
