+++ spatialprofilingtoolbox/control/scripts/guess-channels.py
@@ -4,7 +4,7 @@ import re
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     try:
         import pandas as pd
     except ModuleNotFoundError as e:
