+++ spatialprofilingtoolbox/control/scripts/report-on-logs.py
@@ -14,7 +14,7 @@ import base64
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     try:
         import pandas as pd
         import jinja2
