+++ spatialprofilingtoolbox/db/scripts/modify-constraints.py
@@ -11,7 +11,7 @@ import re
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     try:
         import pandas as pd
     except ModuleNotFoundError as e:
