+++ spatialprofilingtoolbox/workflow/scripts/aggregate-core-results.py
@@ -2,7 +2,7 @@ import argparse
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     from spatialprofilingtoolbox import get_workflow_names
     from spatialprofilingtoolbox import get_workflow
     from spatialprofilingtoolbox import get_integrator
