+++ spatialprofilingtoolbox/workflow/scripts/merge-sqlite-dbs.py
@@ -7,7 +7,7 @@ import argparse
 
 def do_library_imports():
     import spatialprofilingtoolbox
-    from spatialprofilingtoolbox.module_load_error import SuggestExtrasException
+    from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
     try:
         import pandas as pd
     except ModuleNotFoundError as e:
